# docs/co-chatgpt-integration/CO-CHATGPT-INTEGRATION-V1.openapi.yaml
openapi: 3.1.0
info:
  title: Catalyst One — ChatGPT Read-Only Integration V1
  version: 1.0.0
  description: |
    Dedicated read-only integration namespace for ChatGPT executive operational intelligence.
    **GET-only.** No mutation capabilities. Does not expose employee/admin/partner APIs.

    **Channel auth:** `X-ChatGPT-Integration-Key` (server integration key).

    **User auth:** OAuth 2.0 authorization code + PKCE issues a short-lived integration bearer token
    (`aud=catalyst_one_chatgpt`). Employee session JWTs are **not** accepted on integration reads.
servers:
  - url: https://catalyst-one.rupeecatalyst.com
    description: Production (Hostinger)
  - url: http://localhost:3000
    description: Local development
...
security:
  - ChatGptIntegrationApiKey: []
    ChatGptIntegrationOAuth: []
paths:
  /api/integrations/chatgpt/v1/oauth/authorize:
    get:
      tags: [ChatGPT Integration V1]
      summary: OAuth authorization endpoint (authorization code + PKCE)
      operationId: chatgptOAuthAuthorize
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: scope
          in: query
          required: false
          schema:
            type: string
            example: chatgpt:read chatgpt:chanakya
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
      responses:
        "302":
          description: Redirect to Catalyst One consent UI
        "400":
          $ref: "#/components/responses/OAuthError"
        "503":
          $ref: "#/components/responses/NotConfigured"
...
components:
  securitySchemes:
    ChatGptIntegrationApiKey:
      type: apiKey
      in: header
      name: X-ChatGPT-Integration-Key
      description: |
        Dedicated integration channel key (`CHATGPT_INTEGRATION_API_KEY`).
        Required on every integration read alongside the OAuth integration bearer token.
    ChatGptIntegrationOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: /api/integrations/chatgpt/v1/oauth/authorize
          tokenUrl: /api/integrations/chatgpt/v1/oauth/token
          scopes:
            chatgpt:read: Read-only operational intelligence (requires AI_ACCESS + AI_TEXT + AI_CATALYST_INTELLIGENCE)
            chatgpt:chanakya: CHANAKYA / Radar endpoint (requires AI_CHANAKYA)
      description: |
        OAuth 2.0 authorization code + PKCE. Issues a short-lived JWT (`aud=catalyst_one_chatgpt`)
        accepted **only** by `/api/integrations/chatgpt/v1/*`. Employee session JWTs are rejected.
  responses:
    OAuthError:
      description: OAuth request failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    NotConfigured:
      description: Integration key not configured on deployment
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
    ApiError:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message, statusCode]
          properties:
            code:
              type: string
            message:
              type: string
            statusCode:
              type: integer
            correlationId:
              type: string
...